tokyo-covid19-news-flash/pdf-to-text-email/requirements.txt
# Function dependencies, for example:
# package>=version

pdfminer.six
requests
bs4
tabula-py
